+++ src/common/api/profiles.api.ts
@@ -38,3 +38,14 @@ export const savePreferredProfile = (id: string | number, resourceType: string) 
     },
   });
 };
+
+export const deletePreferredProfile = (resourceType: ResourceTypeURL) => {
+  const url = `${PROFILE_PREFERRED_API_ENDPOINT}?resourceType=${encodeURIComponent(resourceType)}`;
+
+  return baseApi.request({
+    url,
+    requestParams: {
+      method: 'DELETE',
+    },
+  });
+};